requirements.txt
streamlit
torch
torchvision
transformers
rasterio
matplotlib
numpy
pillow
albumentations